requirements.txt
# Core dependencies
torch>=2.0.0
torchdiffeq>=0.2.3
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
pyyaml>=6.0
tqdm>=4.62.0

# Machine learning
scikit-learn>=1.0.0
tensorboard>=2.11.0

# Optional dependencies
wandb>=0.13.0
jupyterlab>=3.4.0
pytest>=7.0.0
pytest-cov>=3.0.0

# Visualization
plotly>=5.11.0
networkx>=2.8.0

# Development
black>=22.0.0
flake8>=5.0.0
mypy>=0.990
pre-commit>=2.20.0